requirements.txt
pytest==8.3.5
python-dotenv==1.1.1